Move tail sampling processor to its own package
