+++ cmd/occollector/app/collector/telemetry.go
@@ -29,6 +29,7 @@ import (
 	"github.com/census-instrumentation/opencensus-service/internal/collector/processor"
 	"github.com/census-instrumentation/opencensus-service/internal/collector/processor/nodebatcher"
 	"github.com/census-instrumentation/opencensus-service/internal/collector/processor/queued"
+	"github.com/census-instrumentation/opencensus-service/internal/collector/processor/tail_sampling"
 	"github.com/census-instrumentation/opencensus-service/internal/collector/telemetry"
 )
 
@@ -59,7 +60,7 @@ func initTelemetry(asyncErrorChannel chan<- error, v *viper.Viper, logger *zap.L
 	views = append(views, queued.MetricViews(level)...)
 	views = append(views, nodebatcher.MetricViews(level)...)
 	views = append(views, internal.AllViews...)
-	views = append(views, processor.SamplingProcessorMetricViews(level)...)
+	views = append(views, tailsampling.SamplingProcessorMetricViews(level)...)
 	processMetricsViews := telemetry.NewProcessMetricsViews()
 	views = append(views, processMetricsViews.Views()...)
 	if err := view.Register(views...); err != nil {
